Feature: verify-single with line number when minitest-use-rails is t
minitest-verify-single calls a new function minitest--verify-single-rails when
the minitest-use-rails custom variable is true. this will pass the line number
of the line the point is on to the bin/rails test command. rails version 5 and
above has functionality to run the test for the given line number that is not
built into minitest.

this should solve several issues for rails + minitest users.